requirements.txt
coverage==4.5.1
Django==2.0.6
djangorestframework==3.8.2
django-imagekit==4.0.2
django-leaflet==0.20.0
drf-nested-routers==0.90.2
psycopg2-binary==2.7.5
pillow==5.1.0
raven==6.9.0

# dev
django-debug-toolbar==1.9.1

# This take a lot of time to build, if you need it uncomment and pip install 'em
selenium==3.13.0
